requirements.txt
fire==0.4.0
numpy==1.23.4
pandas==1.5.1
pytorch_lightning==1.7.7
PyYAML==6.0.2
scikit_learn==1.5.1
torch==1.12.1
tqdm==4.64.1
transformers==4.23.1
